Move math dataclasses into models_math.py
Keep the math description data model separate from the high-level biomodel
model. Adds models_base.py (shared VcmlNode / StrEnum base types) and
models_math.py (Constant, MathFunction, MathVariable, subdomains,
equations, stochastic/particle constructs, MathDescription). models.py
re-exports the base types and imports MathDescription for the Application
field; the reader/writer/__init__ import math types from models_math.

No behavior change. make check passes; tests green.
