+++ pyvcell/vcml/__init__.py
@@ -3,31 +3,14 @@ from pyvcell.vcml.field import Field
 from pyvcell.vcml.models import (
     Application,
     Biomodel,
-    Boundaries,
     BoundaryType,
     Compartment,
-    CompartmentSubDomain,
-    Constant,
-    Effect,
     Geometry,
     Image,
-    JumpCondition,
-    JumpProcess,
     Kinetics,
     KineticsParameter,
-    MathBoundaryType,
-    MathDescription,
-    MathFunction,
-    MathVariable,
-    MathVariableType,
-    MembraneSubDomain,
     Model,
     ModelParameter,
-    OdeEquation,
-    ParticleInitialCount,
-    ParticleJumpProcess,
-    ParticleProperties,
-    PdeEquation,
     PixelClass,
     Reaction,
     Simulation,
@@ -38,10 +21,29 @@ from pyvcell.vcml.models import (
     SubVolume,
     SubVolumeType,
     SurfaceClass,
-    VariableInitialCount,
     VCMLDocument,
-    Velocity,
     Version,
+)
+from pyvcell.vcml.models_math import (
+    Boundaries,
+    CompartmentSubDomain,
+    Constant,
+    Effect,
+    JumpCondition,
+    JumpProcess,
+    MathBoundaryType,
+    MathDescription,
+    MathFunction,
+    MathVariable,
+    MathVariableType,
+    MembraneSubDomain,
+    OdeEquation,
+    ParticleInitialCount,
+    ParticleJumpProcess,
+    ParticleProperties,
+    PdeEquation,
+    VariableInitialCount,
+    Velocity,
 )
 from pyvcell.vcml.session import SimulationJob, VCellSession
 from pyvcell.vcml.utils import (
