+++ pyvcell/vcml/models_base.py
@@ -1,0 +1,11 @@
+from enum import Enum
+
+from pydantic import BaseModel
+
+
+class StrEnum(str, Enum):
+    pass
+
+
+class VcmlNode(BaseModel):
+    pass
