+++ pyvcell/vcml/vcml_writer.py
@@ -10,25 +10,27 @@ from pyvcell.vcml.models import (
     Application,
     Biomodel,
     BoundaryType,
-    CompartmentSubDomain,
-    Effect,
     Geometry,
-    JumpCondition,
-    JumpProcess,
     Kinetics,
-    MathBoundaryType,
-    MathDescription,
-    MembraneSubDomain,
     Model,
-    OdeEquation,
-    ParticleJumpProcess,
-    ParticleProperties,
-    PdeEquation,
     Reaction,
     SpeciesMapping,
     SubVolumeType,
     VCMLDocument,
     Version,
+)
+from pyvcell.vcml.models_math import (
+    CompartmentSubDomain,
+    Effect,
+    JumpCondition,
+    JumpProcess,
+    MathBoundaryType,
+    MathDescription,
+    MembraneSubDomain,
+    OdeEquation,
+    ParticleJumpProcess,
+    ParticleProperties,
+    PdeEquation,
 )
